Aim-Backup1203
1. Login screen
· When data is entered in the username field, it is using spell check
and suggesting names. Also it is converting the first char to
uppercase. This should be disabled for user name. - Fixed
· I am assuming that this feature is not enabled for password. -
Confirm if this is true
· Can we use more prominent font and size for the data entered in the
user name field. - Changed
2. Check order status entry
· healthplan dropdown, health plan data field, order id data field -
use more prominent font size and color for the data - Changed
· healthplan field: Left align the data entered - Fixed
· member number - Disable spell check, upper case first letter - Fixed
· Last name - when entering last name field, key board shows 'done'.
However, we may need to enter date of birth! - Date of birth has a
different control and key pad will not be shown.  So, this change is
not required.
· dob: before entering dob, it shows mm/dd/yyy, but after entering the
date, it shows 'nn-nn-nnnn', ie it uses '-' instead of '/'. It should
be consistent. - Fixed
· dob: how can I clear the DOB field. i.e. there should be clear option
in the calendar pop up. -Fixed
3. Member results screen
· Title of the screen shows 'Order Entry'. Should it show, 'Order
Status - Member Results'? - No.  It is basically order inquiry.  So,
this change is not require. - Change the title as 'Order Enquiry -
Member Results'
· When 'back' arrow is pressed, it erases the data entered previously.
It should show the data entered previously. - Fixed
· How do I know that there are more items than displayed on the screen,
and that I can scroll down? AIM has to scroll many times to find the
correct member. So the screen should indicate that there are more
members below. - Scroll bar is not a general practice in iPhone / iPad
app.   We are trying to find other ways to display.
4. order results screen
· How do I know that there are more items than displayed on the screen,
and that I can scroll down? AIM has to scroll many times to find the
correct member. So the screen should indicate that there are more
members below.- Scroll bar is not a general practice in iPhone / iPad
app. We are trying to find other ways to display.
5. order information screen
· How do I know that there are more items than displayed on the screen,
and that I can scroll down? Every order has many notes.- Scroll bar is
not a general practice in iPhone / iPad app. We are trying to find
other ways to display.
· When back arrow is clicked twice, the order results screen transition
happens on the left side, but the member results screen is transitioned
on the right side. It is confusing. It should also transition to the
left side. - Fixed
· Portrait mode: Name of the Ordering provider is displayed in the bar.
But name is not aligned correctly. When you click on the bar the
alignment changes. - Fixed